==> 02_concat_final_csv.ipynb <==
import pandas as pd
import numpy as np

# CONCATTING FINAL CSV FILE

# organized alphabetically
action_comics_vol3 = pd.read_csv("data/action_comics_vol3.csv")
adventures_of_superman_jon_kent = pd.read_csv("data/adventures_of_superman_jon_kent.csv")
aquaman_rebirth = pd.read_csv("data/aquaman_rebirth.csv")
batman_and_robin = pd.read_csv("data/batman_and_robin.csv")
batman_gotham_knights = pd.read_csv("data/batman_gotham_knights.csv")
batman_vol1 = pd.read_csv("data/batman_vol1.csv")
batman_vol2 = pd.read_csv("data/batman_vol2.csv")
batman_vol3 = pd.read_csv("data/batman_vol3.csv")
detective_comics_vol1 = pd.read_csv("data/detective_comics_vol1.csv")
detective_comics_vol2 = pd.read_csv("data/detective_comics_vol2.csv")
detective_comics_vol3 = pd.read_csv("data/detective_comics_vol3.csv")
gen_13 = pd.read_csv("data/gen_13.csv")
gotham_academy = pd.read_csv("data/gotham_academy.csv")
grayson = pd.read_csv("data/grayson.csv")
green_arrow_vol2 = pd.read_csv('data/green_arrow_vol2.csv')
green_lanturn_vol2 = pd.read_csv("data/green_lanturn_vol2.csv")
harley_quinn_vol2 = pd.read_csv("data/harley_quinn_vol2.csv")
harley_quinn_vol3 = pd.read_csv("data/harley_quinn_vol3.csv")
hellblazer = pd.read_csv("data/hellblazer.csv")
house_of_mystery_pt1 = pd.read_csv('data/house_of_mystery_pt1.csv')
house_of_mystery_pt2 = pd.read_csv('data/house_of_mystery_pt2.csv')
house_of_secrets = pd.read_csv('data/house_of_secrets.csv')
JLA = pd.read_csv("data/JLA.csv")
justice_league_dark = pd.read_csv("data/justice_league_dark.csv")
nightwing_vol2 = pd.read_csv('data/nightwing_vol2.csv')
nightwing_vol4 = pd.read_csv('data/nightwing_vol4.csv')
red_hood_and_arsenal = pd.read_csv('data/red_hood_and_arsenal.csv')
red_robin = pd.read_csv("data/red_robin.csv")
robin_son_of_batman = pd.read_csv("data/robin_son_of_batman.csv")
robin1993_2009 = pd.read_csv("data/robin1993_2009.csv")
super_sons = pd.read_csv("data/super_sons.csv")
superboy = pd.read_csv("data/superboy.csv")
superman_son_of_kal_el = pd.read_csv("data/superman_son_of_kal_el.csv")
superman_vol1_pt1 = pd.read_csv("data/superman_vol1_pt1.csv")
superman_vol1_pt2 = pd.read_csv("data/superman_vol1_pt2.csv")
superman_vol2 = pd.read_csv("data/superman_vol2.csv")
superman_vol3 = pd.read_csv("data/superman_vol3.csv")
superman_vol4 = pd.read_csv("data/superman_vol4.csv")
superman_vol5 = pd.read_csv("data/superman_vol5.csv")
swamp_thing_vol2 = pd.read_csv("data/swamp_thing_vol2.csv")
swamp_thing_vol3 = pd.read_csv("data/swamp_thing_vol3.csv")
swamp_thing_vol4 = pd.read_csv("data/swamp_thing_vol4.csv")
teen_titans1966_1996 = pd.read_csv("data/teen_titans1966_1996.csv")
teen_titans2014 = pd.read_csv("data/teen_titans2014.csv")
the_books_of_magic = pd.read_csv("data/the_books_of_magic.csv")
the_flash_vol1 = pd.read_csv("data/the_flash_vol1.csv")
the_flash_vol2 = pd.read_csv("data/the_flash_vol2.csv")
the_flash_vol3 = pd.read_csv("data/the_flash_vol3.csv")
the_sandman_vol2 = pd.read_csv("data/the_sandman_vol2.csv")
we_are_robin = pd.read_csv("data/we_are_robin.csv")
wonder_woman_vol2 = pd.read_csv("data/wonder_woman_vol2.csv")
wonder_woman_vol3 = pd.read_csv("data/wonder_woman_vol3.csv")
wonder_woman_vol4 = pd.read_csv("data/wonder_woman_vol4.csv")
wonder_woman_vol5 = pd.read_csv("data/wonder_woman_vol5.csv")
young_justice = pd.read_csv("data/young_justice.csv")


df = pd.concat([action_comics_vol3, 
                adventures_of_superman_jon_kent,
                aquaman_rebirth, 
                batman_and_robin,
                batman_gotham_knights,
                batman_vol1, 
                batman_vol2, 
                batman_vol3,
                detective_comics_vol1, 
                detective_comics_vol2, 
                detective_comics_vol3,
                gen_13,
                gotham_academy, 
                grayson,
                green_arrow_vol2,
                green_lanturn_vol2,
                harley_quinn_vol2,
                harley_quinn_vol3,
                hellblazer, 
                house_of_mystery_pt1,
                house_of_mystery_pt2,
                house_of_secrets,
                JLA, 
                justice_league_dark, 
                nightwing_vol2,
                nightwing_vol4,
                red_hood_and_arsenal,
                red_robin, 
                robin_son_of_batman,
                robin1993_2009, 
                super_sons, 
                superboy, 
                superman_son_of_kal_el,
                superman_vol1_pt1, 
                superman_vol1_pt2, 
                superman_vol2, 
                superman_vol3, 
                superman_vol4, 
                superman_vol5, 
                swamp_thing_vol2, 
                swamp_thing_vol3, 
                swamp_thing_vol4,
                teen_titans1966_1996, 
                teen_titans2014,
                the_books_of_magic,
                the_flash_vol1, 
                the_flash_vol2, 
                the_flash_vol3, 
                the_sandman_vol2,
                we_are_robin, 
                wonder_woman_vol2,
                wonder_woman_vol3, 
                wonder_woman_vol4, 
                wonder_woman_vol5,
                young_justice
               ], ignore_index=True)

#use when sometimes adding one new csv file
#df= pd.concat([final, sample], ignore_index=True)

#df.to_csv("final_dataset.csv", index=False)

df.to_csv("data/final/final_dataset.csv", index=False)
